Add a new Scheduler for the Selenium Server
The idea is to finally get to a point where the
standalone server is _actually_ just a Grid of
one node, though it'll be possible to add new
Nodes at runtime.

The other thing we're attempting to do here is
to create a code base that's a little easier to
read and reason about --- the current Grid code
is notoriously hard to grok.

The key moving pieces are:

`Host`: represents a physical machine that is
  capable of running multiple sessions.

`Distributor`: Responsible for ensuring an even
  distribution of sessions between hosts.

`Scheduler`: The entry point that should be
  used for actually getting a new `ActiveSession`.

The `Scheduler` handles things like attempting to
recreate a session should the `SessionFactory`
fail, as well as providing a mechanism to
reschedule an attempt to create a session.

The next step will be to make this "pin compatible"
with the existing `NewSessionPipeline` and hide it
behind a flag so we can try using this in production.

Exciting stuff, I think we can all agree.